Move discovery functions to their own module
