+++ tractor/__init__.py
@@ -12,9 +12,8 @@ from trio import MultiError
 
 from . import log
 from ._ipc import _connect_chan, Channel, Context
-from ._actor import (
-    Actor, _start_actor, Arbiter, get_arbiter, find_actor, wait_for_actor
-)
+from ._discovery import get_arbiter, find_actor, wait_for_actor
+from ._actor import Actor, _start_actor, Arbiter
 from ._trionics import open_nursery
 from ._state import current_actor
 from ._exceptions import RemoteActorError, ModuleNotExposed
